==> 2_Train_GMP_with_data_generator_edited_model.ipynb <==
We will also use the previously created train and validation files.

# import the necessary packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import tensorflow as tf
import os
from scipy import ndimage
from sklearn.model_selection import train_test_split

tf.config.list_physical_devices("GPU")

sub_window_and_count_path_train = "../../S_lab_TasselNet/Method_2/final_train_sub_windows_and_counts"
sub_window_and_count_path_valid = "../../S_lab_TasselNet/Method_2/final_valid_sub_windows_and_counts"

%%time
# load the data

# subwindow data
X_train = np.load(os.path.join(sub_window_and_count_path_train, 'train_sub_windows.npy'))
print(X_train.shape)

# sub counts data
y_train = np.load(os.path.join(sub_window_and_count_path_train,'train_sub_counts.npy'))
print(y_train.shape)

%%time
# load validation data
# sub windows
X_valid = np.load(os.path.join(sub_window_and_count_path_valid,'valid_sub_windows.npy'))
print(X_valid.shape)

# sub counts data
y_valid = np.load(os.path.join(sub_window_and_count_path_valid,'valid_sub_counts.npy'))
print(y_valid.shape)

# load and add the new layers to the prev trained model
gmp_model = tf.keras.models.load_model('../../S_lab_TasselNet/models/tasselnet_global_max_pool_04_23_2022.h5')
# look at the input shape
gmp_model.input

gmp_model.summary()

# need to remove some of the last layers
output_trial = gmp_model.layers[-7].output

# create the functional API model
reduced_model = tf.keras.models.Model(gmp_model.input, output_trial)
reduced_model.summary()

# okay, now need to add back the dropout, the dense and activation

# add dropout
added_dropout = tf.keras.layers.Dropout(0.5, name = "New_dropout_0")(gmp_model.layers[-7].output)

# add global max pooling
added_flatten = tf.keras.layers.GlobalMaxPooling2D()(added_dropout)

# add dense
added_dense = tf.keras.layers.Dense(64, name = "New_Dense_0")(added_flatten)

# add activation
added_Act = tf.keras.layers.Activation('relu', name = "New_Activation_0")(added_dense)

# add dropout
added_dropout2 = tf.keras.layers.Dropout(0.5, name = "New_dropout_1")(added_Act)

# add another dense
added_dense_1 = tf.keras.layers.Dense(32, name = "New_Dense_1")(added_dropout2)

# add activation
added_Act_1 = tf.keras.layers.Activation('relu', name = "New_Activation_1")(added_dense_1)

# add dropout
added_dropout3 = tf.keras.layers.Dropout(0.5, name = "New_dropout_2")(added_Act_1)

# add pred head
added_dense_3 = tf.keras.layers.Dense(1, name = "New_Dense_2")(added_dropout3)

# add activation
added_Act_2 = tf.keras.layers.Activation('relu', name = "New_Activation_2")(added_dense_3)

# define the new model with functional API
new_model = tf.keras.models.Model(gmp_model.input, added_Act_2)

# print the model summary
new_model.summary()

This is where we need to write the function - Rylie has two functions a generator, and also a batch-wise prediction function. May be we only need the generator function? Let's see.

def data_generator(x_data, y_data, batch_size, shuffle=False, peek=False, verbose=False):
    num_samples = len(x_data)
    indices = np.arange(num_samples)
    
    if peek:    # Give first batch unshuffled and don't change start index when peeking for training
        end = min(batch_size, num_samples)
        if verbose:
            print(f"Generating peeking batch up to index {end}")
        batch_x = x_data[:end]
        batch_y = y_data[:end]
        peek = False
        yield (batch_x, batch_y)

    while True:    # Loop indefinitely for epochs
        # Shuffle indices at the start of each epoch after the peek, if shuffle is enabled
        if shuffle:
            np.random.shuffle(indices)
        
        for start in range(0, num_samples, batch_size):
            end = min(start + batch_size, num_samples)
            batch_indices = indices[start:end]

            # Print batch indices if verbose
            if verbose:
                # Warning: calling verbose when shuffling will usually clutter output
                if shuffle:
                    if len(batch_indices) < 16:
                        print(f"Batch indices: {np.sort(batch_indices)}")
                    else:
                        print(f"Printing batch indices would clutter output. Skipped.")
                    print(f"Length: {len(batch_indices)}")
                else:
                    print(f"Generating batch from index {start} to {end}")
                # traceback.print_stack()

            # Generate batches
            batch_x = x_data[batch_indices]
            batch_y = y_data[batch_indices]

            # Yield the current batch
            yield (batch_x, batch_y)

%%time
batch_size = 1024
generator_batch_size = 1024
train_gen = data_generator(X_train, y_train, batch_size, shuffle=False, peek=True, verbose=False)
val_gen = data_generator(X_valid, y_valid, generator_batch_size, shuffle=False, peek=True, verbose=False)

steps_per_epoch = int(np.ceil(len(X_train) / batch_size))
train_steps = int(np.ceil(len(X_train) / generator_batch_size))
validation_steps = int(np.ceil(len(X_valid) / generator_batch_size))

steps_per_epoch, train_steps, validation_steps

%%time
# frozen model

#     freeze layers
for layer in gmp_model.layers:
    layer.trainable = False
    
# compile the model
opt = tf.keras.optimizers.Adam(learning_rate=0.0001)
new_model.compile(loss='mean_squared_error', optimizer=opt, metrics = ['mean_absolute_error'])
    
# add early stopping
es = tf.keras.callbacks.EarlyStopping(monitor='val_loss', restore_best_weights = True, verbose=1, patience=10)

# fit the generator - the steps are a must here
history = new_model.fit(train_gen,
          validation_data = val_gen, steps_per_epoch=steps_per_epoch, 
                                  validation_steps=validation_steps,                        
          epochs = 100, callbacks = [es])

%%time
# # unfreeze few layers and retrain - finetune the model
gmp_model.trainable = True

set_trainable = False

for layer in gmp_model.layers:
    if layer.name == 'conv2d_3':
        set_trainable = True
    if set_trainable:
        layer.trainable = True
    else:
        layer.trainable = False
            
    # # compile the mdoel
opt = tf.keras.optimizers.Adam(learning_rate=0.0001)
new_model.compile(loss='mean_squared_error', optimizer=opt, metrics = ['mean_absolute_error'])
    
    # # add early stopping
es = tf.keras.callbacks.EarlyStopping(monitor='val_loss', restore_best_weights = True, verbose=1, patience=10)
    
    # # fit the model (fine tuned)
history_new = new_model.fit(train_gen,
          validation_data = val_gen, steps_per_epoch=steps_per_epoch, 
                                  validation_steps=validation_steps,                        
          epochs = 100, callbacks = [es])


# save the finetuned model
model_name = 'trained_gmp_model_dense_32.h5'
new_model.save('models' + '/' + model_name)
