==> 3_Extract_features_model_rest_of_the_blocks.ipynb <==
import tensorflow as tf
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import os

# load the trained model
gmp_model = tf.keras.models.load_model('models/trained_gmp_model_dense_32_new.h5')

gmp_model.summary()

# check gpu utility
tf.config.list_physical_devices("GPU")

# feature extractor input
feat_ext_input = gmp_model.input

# feature extractor output - do this at the ReLu activation layer - as this will give the same features as the dropout layer (It does not matter if it is the dropout or the activation layer, the extracted features will be the same)
feat_ext_output = gmp_model.layers[-4].output

feature_extractor_model = tf.keras.models.Model(inputs = feat_ext_input, outputs = feat_ext_output)

feature_extractor_model.summary()

# We might need to be cautious here though, as dropout will set some neurons to 0, and the extracted features will therefore become zero - workaround - use the Dense layer to extract features before the pred head (and the dropout layer)

def extract_features_and_get_counts(model, preprocessed_path, all_in_folder, digits_image, digits_dense, character, store_path):
    necessary_sub_images = [item for item in all_in_folder if item.split(".")[0][-digits_image:] == ('_' + character)]
    necessary_sub_images.sort()
    necessary_sub_dense_maps = [item for item in all_in_folder if item.split(".")[0][-digits_dense:] == ('_' + character + '_density')]
    necessary_sub_dense_maps.sort()
    print(necessary_sub_images[0])
    # check if the image and density maps tally
    print(np.mean([item.split('.')[0] for item in necessary_sub_images] == [item.split(".")[0][:-8] for item in necessary_sub_dense_maps]))
    
    collect_extracted_features = []
    collect_corresponding_counts = []

    # collect the subwindow-wise features and corresponding counts
    for i in range(len(necessary_sub_images)):
        # load the subimage
        load_image = np.load(os.path.join(preprocessed_path, necessary_sub_images[i]))
        # reshape it to get the prediction
        load_image_reshape = load_image[None, ...]
        # get the prediction from the model
        extracted_features = model.predict(load_image_reshape, verbose = 0)
        collect_extracted_features.append(extracted_features)

        # load the count map
        load_count_map = np.load(os.path.join(preprocessed_path, necessary_sub_dense_maps[i]))
        count = np.sum(load_count_map)
        collect_corresponding_counts.append(count)

    # collect all extracted features data and create a dataframe
    Features_data = pd.DataFrame(np.vstack(collect_extracted_features), columns = ['feature_' + str(i) for i in range(np.vstack(collect_extracted_features).shape[1])])
    # dataframe for counts
    counts_subwindow = pd.DataFrame(collect_corresponding_counts, columns = ['tassel_count'])
    # combine the two dataframes
    combined_df = pd.concat((Features_data,counts_subwindow), axis = 1 )
    # save this dataframe
    save_path_name = store_path + 'extracted_features_sub_window_' + character + '.csv'
    combined_df.to_csv(save_path_name, index = False)

    return(combined_df)

# There's actually another thing to check before we close off this notebook -  we need to make sure the sums of density maps match with what we had stored earlier at the location all_preprocessed_data/Block_0103/sub_count_dfs/ - Do this in a separate notebook?

Block 0105

prev_preprocessed_image_loc_0105 = "all_preprocessed_data/Block_0105/sub_images_and_counts"

folder_contents_0105 = os.listdir(prev_preprocessed_image_loc_0105)
folder_contents_0105.sort()

# folder_contents

folder_contents_0105[0]

s_path_0105 = 'all_preprocessed_data/Block_0105//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0105 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0105, folder_contents_0105, 2, 10, character, s_path_0105)
    catch_all_dfs_0105.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0105 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0105, folder_contents_0105, 3, 11, character, s_path_0105)
    catch_all_dfs_0105.append(features_df)

# catch_all_dfs_0105

Block 0106

prev_preprocessed_image_loc_0106 = "all_preprocessed_data/Block_0106/sub_images_and_counts"

folder_contents_0106 = os.listdir(prev_preprocessed_image_loc_0106)
folder_contents_0106.sort()

# folder_contents

folder_contents_0106[0]

s_path_0106 = 'all_preprocessed_data/Block_0106//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0106 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0106, folder_contents_0106, 2, 10, character, s_path_0106)
    catch_all_dfs_0106.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0106 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0106, folder_contents_0106, 3, 11, character, s_path_0106)
    catch_all_dfs_0106.append(features_df)

# catch_all_dfs_0105

Block 0201

prev_preprocessed_image_loc_0201 = "all_preprocessed_data/Block_0201/sub_images_and_counts"

folder_contents_0201 = os.listdir(prev_preprocessed_image_loc_0201)
folder_contents_0201.sort()

# folder_contents_0201

folder_contents_0201[0]

s_path_0201 = 'all_preprocessed_data/Block_0201//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0201 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0201, folder_contents_0201, 2, 10, character, s_path_0201)
    catch_all_dfs_0201.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0201 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0201, folder_contents_0201, 3, 11, character, s_path_0201)
    catch_all_dfs_0201.append(features_df)

# catch_all_dfs_0105

Block 0202

prev_preprocessed_image_loc_0202 = "all_preprocessed_data/Block_0202/sub_images_and_counts"

folder_contents_0202 = os.listdir(prev_preprocessed_image_loc_0202)
folder_contents_0202.sort()

# folder_contents_0201

folder_contents_0202[0]

s_path_0202 = 'all_preprocessed_data/Block_0202//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0202 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0202, folder_contents_0202, 2, 10, character, s_path_0202)
    catch_all_dfs_0202.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0202 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0202, folder_contents_0202, 3, 11, character, s_path_0202)
    catch_all_dfs_0202.append(features_df)

# catch_all_dfs_0105

Block 0205

prev_preprocessed_image_loc_0205 = "all_preprocessed_data/Block_0205/sub_images_and_counts"

folder_contents_0205 = os.listdir(prev_preprocessed_image_loc_0205)
folder_contents_0205.sort()

# folder_contents_0201

folder_contents_0205[0]

s_path_0205 = 'all_preprocessed_data/Block_0205//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0205 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0205, folder_contents_0205, 2, 10, character, s_path_0205)
    catch_all_dfs_0205.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0205 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0205, folder_contents_0205, 3, 11, character, s_path_0205)
    catch_all_dfs_0205.append(features_df)

Block 0206

prev_preprocessed_image_loc_0206 = "all_preprocessed_data/Block_0206/sub_images_and_counts"

folder_contents_0206 = os.listdir(prev_preprocessed_image_loc_0206)
folder_contents_0206.sort()

# folder_contents_0201

folder_contents_0206[0]

s_path_0206 = 'all_preprocessed_data/Block_0206//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0206 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0206, folder_contents_0206, 2, 10, character, s_path_0206)
    catch_all_dfs_0206.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0206 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0206, folder_contents_0206, 3, 11, character, s_path_0206)
    catch_all_dfs_0206.append(features_df)

Block 0302

prev_preprocessed_image_loc_0302 = "all_preprocessed_data/Block_0302/sub_images_and_counts"

folder_contents_0302 = os.listdir(prev_preprocessed_image_loc_0302)
folder_contents_0302.sort()

# folder_contents_0201

folder_contents_0302[0]

s_path_0302 = 'all_preprocessed_data/Block_0302//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0302 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0302, folder_contents_0302, 2, 10, character, s_path_0302)
    catch_all_dfs_0302.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0302 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0302, folder_contents_0302, 3, 11, character, s_path_0302)
    catch_all_dfs_0302.append(features_df)

Block 0303

prev_preprocessed_image_loc_0303 = "all_preprocessed_data/Block_0303/sub_images_and_counts"

folder_contents_0303 = os.listdir(prev_preprocessed_image_loc_0303)
folder_contents_0303.sort()

# folder_contents_0201

folder_contents_0303[0]

s_path_0303 = 'all_preprocessed_data/Block_0303//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0303 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0303, folder_contents_0303, 2, 10, character, s_path_0303)
    catch_all_dfs_0303.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0303 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0303, folder_contents_0303, 3, 11, character, s_path_0303)
    catch_all_dfs_0303.append(features_df)

Block 0304

prev_preprocessed_image_loc_0304 = "all_preprocessed_data/Block_0304/sub_images_and_counts"

folder_contents_0304 = os.listdir(prev_preprocessed_image_loc_0304)
folder_contents_0304.sort()

# folder_contents_0201

folder_contents_0304[0]

s_path_0304 = 'all_preprocessed_data/Block_0304//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0304 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0304, folder_contents_0304, 2, 10, character, s_path_0304)
    catch_all_dfs_0304.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0304 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0304, folder_contents_0304, 3, 11, character, s_path_0304)
    catch_all_dfs_0304.append(features_df)

Block 0305

prev_preprocessed_image_loc_0305 = "all_preprocessed_data/Block_0305/sub_images_and_counts"

folder_contents_0305 = os.listdir(prev_preprocessed_image_loc_0305)
folder_contents_0305.sort()

# folder_contents_0201

folder_contents_0305[0]

s_path_0305 = 'all_preprocessed_data/Block_0305//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0305 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0305, folder_contents_0305, 2, 10, character, s_path_0305)
    catch_all_dfs_0305.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0305 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0305, folder_contents_0305, 3, 11, character, s_path_0305)
    catch_all_dfs_0305.append(features_df)

Block 0306

prev_preprocessed_image_loc_0306 = "all_preprocessed_data/Block_0306/sub_images_and_counts"

folder_contents_0306 = os.listdir(prev_preprocessed_image_loc_0306)
folder_contents_0306.sort()

# folder_contents_0201

folder_contents_0306[0]

s_path_0306 = 'all_preprocessed_data/Block_0306//TS_ready_data_frames/'

# get the predictions for all subimages for the block? do separately for single digit subimages and double digit subimages separately
character_list = list(range(10))
character_list = list(map(str, character_list))

%%time

# for single digit subwindows
catch_all_dfs_0306 = []
for character in character_list:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0306, folder_contents_0306, 2, 10, character, s_path_0306)
    catch_all_dfs_0306.append(features_df)

# get the predictions for double digits
character_list_double = list(range(10, 12))
character_list_double = list(map(str, character_list_double))

%%time

# for double digit subwindows
catch_all_dfs_0306 = []
for character in character_list_double:
    features_df = extract_features_and_get_counts(feature_extractor_model, prev_preprocessed_image_loc_0306, folder_contents_0306, 3, 11, character, s_path_0306)
    catch_all_dfs_0306.append(features_df)
