==> 4_verify_the_counts.ipynb <==
# we need to verify the counts we have computed for the sub density maps are the same from the notebooks 1 and 3.

import pandas as pd
import numpy as np
import os

file_loc_preproc_counts = "all_preprocessed_data/Block_0103/sub_count_dfs/"
file_loc_preproc_counts

contents_this_folder = os.listdir(file_loc_preproc_counts)
contents_this_folder.sort()

contents_this_folder = contents_this_folder[1:] # as there was a pynb checkpointing folder

contents_this_folder

# contents_this_folder.sort()

%%time
catch_all_dfs = []
for file in contents_this_folder:
    df_path = os.path.join(file_loc_preproc_counts, file)
    df_read = pd.read_csv(df_path)
    catch_all_dfs.append(df_read)

len(catch_all_dfs)

# concatenate all these data together
all_preproc_density_df = pd.concat((catch_all_dfs[0], catch_all_dfs[1][['sub_counts']], catch_all_dfs[12][['sub_counts']], catch_all_dfs[13][['sub_counts']]
          , catch_all_dfs[14][['sub_counts']], catch_all_dfs[15][['sub_counts']], catch_all_dfs[16][['sub_counts']], catch_all_dfs[17][['sub_counts']],
           catch_all_dfs[18][['sub_counts']], catch_all_dfs[19][['sub_counts']], catch_all_dfs[2][['sub_counts']], catch_all_dfs[3][['sub_counts']], 
           catch_all_dfs[4][['sub_counts']], catch_all_dfs[5][['sub_counts']], 
           catch_all_dfs[6][['sub_counts']], catch_all_dfs[7][['sub_counts']], catch_all_dfs[8][['sub_counts']], catch_all_dfs[9][['sub_counts']],
           catch_all_dfs[10][['sub_counts']], catch_all_dfs[11][['sub_counts']]), axis = 1)

all_preproc_density_df.shape

all_preproc_density_df.columns = ['image_context'] + ['sub_count_' + str(i) for i in range(len(catch_all_dfs))]

all_preproc_density_df

final_ts_dfs_path = "all_preprocessed_data/Block_0103/TS_ready_data_frames/"

contents_final_df_path = os.listdir(final_ts_dfs_path)
contents_final_df_path.sort()

contents_final_df_path = contents_final_df_path[1:]

contents_final_df_path

late_dates = ['extracted_features_sub_window_10.csv', 'extracted_features_sub_window_11.csv']

early_dates = [i for i in contents_final_df_path if i not in late_dates]

final_dates = early_dates + late_dates

final_dates

%%time
catch_final_dfs = []
for file in final_dates:
    df_path = os.path.join(final_ts_dfs_path, file)
    df_read = pd.read_csv(df_path)
    df_count = df_read[['tassel_count']]
    catch_final_dfs.append(df_count)

counts_only_df = pd.DataFrame(np.hstack(catch_final_dfs).T, columns = ['sub_count_' + str(i) for i in range(20)])

counts_only_df

sub_win_name = pd.DataFrame(final_dates, columns = ['Sub_window_name'])

final_df_facts = pd.concat((sub_win_name, counts_only_df), axis = 1)

final_df_facts.head()

final_df_facts['Sub_window_name'] = [int(item.split(".")[0].split("_")[-1]) for item in final_df_facts['Sub_window_name']]

final_df_facts.head()

final_df_facts = final_df_facts.sort_values(['Sub_window_name']).reset_index(drop = True)

final_df_facts

# see if the two dataframes are teh same
np.mean(final_df_facts.iloc[:,1:] == all_preproc_density_df.iloc[:,1:])

final_df_facts.iloc[:,1:] == all_preproc_density_df.iloc[:,1:]
