==> 7_comps_Computing_image_wise_forecasts_and_metrics_block_0103.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import os
from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
from scipy.stats import pearsonr

# path to dataframes
path_to_dfs = "Reproduce_comps/forecasted_counts/block_0103/"

# os.listdir(path_to_dfs)

averaged_forecasts_csvs = []
for file in os.listdir(path_to_dfs):
    if file[:8] =='averaged':
        averaged_forecasts_csvs.append(file)

averaged_forecasts_csvs.sort()
averaged_forecasts_csvs

# add the 10, 11 at the end
csv_files_10_11 = ['averaged_forecasts_sub_10.csv', 'averaged_forecasts_sub_11.csv']

other_files = [i for i in averaged_forecasts_csvs if i not in csv_files_10_11]

ordered_csv_files = other_files + csv_files_10_11

ordered_csv_files

all_dfs = []
for file in ordered_csv_files:
    read_df = pd.read_csv(path_to_dfs + '/' + file)
    all_dfs.append(read_df)

# make a single dataframe out of the many
combined_df = pd.concat(all_dfs, axis = 1)

combined_df.head()

combined_df.shape

# get the true values
True_values_df = combined_df[['True_value']]

True_values_df

# sum the true values - since the sub images are not overlapping, we can get the sum across columns in the above dataframe
sum_true_values = True_values_df.sum(axis = 1)

sum_true_values

# get the forecasted values
Forecasted_values_df = combined_df[['Forecasted_value']]

Forecasted_values_df

Forecasted_values_df.shape

sum_forecasted_values = Forecasted_values_df.sum(axis = 1)

sum_forecasted_values

# concatenate the sum dataframes
block_0103_true_and_forecasted_values_df = pd.concat((sum_true_values, sum_forecasted_values), axis = 1)

block_0103_true_and_forecasted_values_df.columns = ["True_count", "Forecasted_count"]

block_0103_true_and_forecasted_values_df

block_0103_true_and_forecasted_values_df.to_csv("Reproduce_comps/final_forecasted_counts/block_0103_finals_forecasts.csv", index = False)

# # well we need to verify that the above true values are indeed correct
# from preprocess script we have the following counts
# [43, 49, 53, 59, 45, 42, 34, 39, 37, 43, 41, 39, 43, 40, 39, 41, 31, 32, 40, 29] # the last seven numbers match with what we have.

# blockwise mean squared error
rmse = np.sqrt(mean_squared_error(sum_true_values, sum_forecasted_values))
rmse

# blockwise mean absolute error
mae = mean_absolute_error(sum_true_values, sum_forecasted_values)
mae

corr = pearsonr(sum_true_values, sum_forecasted_values)
corr

r2 = r2_score(sum_true_values, sum_forecasted_values)
r2

# May be define a function for this, so that it will be easier to get the forecasted dataframes for all blocks

# we need first the path to dfs per block
def get_final_forecasted_dfs(path_to_dfs_in_block, block_name, true_values_col_name, forecasted_values_col_name, forecast_path):
    # get the csv files that have the averaged 
    average_frcts_csv_files = [file for file in os.listdir(path_to_dfs_in_block) if file[:8] == 'averaged']
    # sort these files
    average_frcts_csv_files.sort()
    # get the later images to the end of the list
    csv_files_10_11 = ['averaged_forecasts_sub_10.csv', 'averaged_forecasts_sub_11.csv']
    # remove these from the total list
    other_files = [i for i in average_frcts_csv_files if i not in csv_files_10_11]
    # add the csv files in order
    ordered_csv_files = other_files + csv_files_10_11
    # print the ordered list of csv files
    print(ordered_csv_files)
    # read and append the list of the dfs
    all_dfs = [pd.read_csv(path_to_dfs_in_block + '/' + df) for df in ordered_csv_files]
    # combine all these dfs together
    combined_df = pd.concat(all_dfs, axis = 1)
    print(combined_df.shape)
    # extract the true value columns only across the sub-images
    True_counts_df = combined_df[[true_values_col_name]]
    # sum the true values dfs
    total_true_values = True_counts_df.sum(axis = 1)
    # print these true values for later comparisons
    print(total_true_values)
    # extract the true value columns across sub images
    Forecasted_counts_df = combined_df[[forecasted_values_col_name]]
    # sum the forecasted values
    total_forecasted_values = Forecasted_counts_df.sum(axis = 1)
    # concatenate the sum dataframes
    true_and_forecasted_values_df = pd.concat((total_true_values, total_forecasted_values), axis = 1)
    # add column titles to the df
    true_and_forecasted_values_df.columns = ["True_count", "Forecasted_count"]
    # save this file
    file_name = forecast_path + '/' + block_name + '.csv'
    true_and_forecasted_values_df.to_csv(file_name, index = False)
    # blockwise mean squared error
    rmse = np.sqrt(mean_squared_error(total_true_values, total_forecasted_values))
    # blockwise mae
    mae = mean_absolute_error(total_true_values, total_forecasted_values)

    return(true_and_forecasted_values_df, rmse, mae,ordered_csv_files)

# see if the function works
df, rmse_0103, mae_0103, ordered_files_0103 = get_final_forecasted_dfs(path_to_dfs, 'block_0103', 'True_value', 'Forecasted_value', 'final_forecasted_counts')

df

rmse_0103

mae_0103

ordered_files_0103

import numpy as np
import pandas as pd
import os

# Getting the coverages and the widths for the forecasted values - This can be done with the saved all forecasts npy files
# location for the forecast files (all forecasts)
location_all_forecasts = 'Reproduce_comps/forecasted_counts/block_0103'

# contents at this location
all_contents = os.listdir(location_all_forecasts)
all_contents.sort()

# we need the npy files for all forecasts
all_forecast_files = [file for file in all_contents if file[:3] == 'all']

# arange the files in order
later_npy_files = ['all_forecasts_sub_10.npy', 'all_forecasts_sub_11.npy']
first_files = [file for file in all_forecast_files if file not in later_npy_files]

final_all_forecast_files = first_files + later_npy_files

loaded_npy_files = []
for file in final_all_forecast_files:
    joined_path = os.path.join(location_all_forecasts, file)
    load_file = np.load(joined_path)
    loaded_npy_files.append(load_file)

len(loaded_npy_files)

# 5.3671412 + 3.80887


# loaded_npy_files[0] 

# loaded_npy_files[1] 

# try_0 = loaded_npy_files[0] + loaded_npy_files[1]

output = sum(loaded_npy_files)

output.shape

final_array = output.reshape(4000,7)

final_array[0,:].shape

li_train = np.percentile(final_array, axis = 0, q = (2.5, 97.5))[0,:].reshape(-1,1)    
ui_train = np.percentile(final_array, axis = 0, q = (2.5, 97.5))[1,:].reshape(-1,1)

li_train.shape

ui_train.shape

width_train = ui_train - li_train
avg_width_train = width_train.mean(0)[0]

avg_width_train

y_traina = block_0103_true_and_forecasted_values_df[["True_count"]].values

y_traina

ind_train = (y_traina >= li_train) & (y_traina <= ui_train)
coverage_train= ind_train.mean(0)[0]

coverage_train

Verify the true counts we have for the test data are indeed correct

# Verify the true counts - from np arrays - location Block_0103/sub_images_and_counts

sub_count_loc = 'all_preprocessed_data/Block_0103/sub_images_and_counts'

sub_density_maps = [i for i in os.listdir(sub_count_loc) if i.split(".")[0][-7:] == 'density']
sub_density_maps.sort()

# get the test dates
test_time_periods = ['2020_08_26', '2020_08_27', '2020_08_28' ,'2020_08_31', '2020_09_02', '2020_09_07', '2020_09_16']

test_time_periods = ['Block0103_' + i for i in test_time_periods]

# print the test dates
test_time_periods

# get the density maps for these days only for the computation of the total true counts of the test images
test_density_maps = [i for i in sub_density_maps if i[:20] in test_time_periods]
test_density_maps.sort()

%%time
# get the true counts
true_counts_in_order = []
step = 0
# we have only seven time points
for u in range(7):
    catch_counts = []
    # for each time steps we have 12 images, and since the sub windows are not overlapping we can add the values straightaway eben without sorting
    for j in range(step, step + 12):
        total_count = np.sum(np.load(os.path.join(sub_count_loc, test_density_maps[j])))
        catch_counts.append(total_count)
    true_counts_in_order.append(np.sum(catch_counts))
    step = step + 12

true_counts_in_order

df[['True_count']].values.flatten()

# see if the two values we get match 

np.mean(np.round(true_counts_in_order, 0) == np.round(df[['True_count']].values.flatten(), 0))
